+++ src/hooks/use-top-articles.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import axios from "axios";
 
 import { ArticlesResponse } from '@/types/articlesResponse';
-import { Article } from "@/types/articleSection";
+import { Article } from "@/types/Article";
 
 export const useTopArticles = () => {
   const [articles, setArticles] = useState<Article[]>([]);
