+++ src/models/Article.tsx
@@ -1,26 +1,13 @@
 import mongoose, { Schema, Document } from "mongoose";
 
-// Defining the interface for the Article model
-export interface Article extends Document {
-  id: string;
-  title: string;
-  content?: string;
-  publishedIn: Date;
-  featuredImage?: string; // Path to the featured image
-  publisherID: string; // Reference to User ID
-  voteCount?: number;
-  postTags: string[];
-  modifiedIn?: Date;
-  trashed?: boolean;
-  category: string;
-  author: string;
-}
+import Article from "@/types/Article";
 
 // Defining the Article schema
 const ArticleSchema: Schema<Article> = new Schema(
   {
     id: { type: String, required: true, trim: true, unique: true }, // Reference to the wordpressd database for this
     title: { type: String, required: true, trim: true },
+    oneLiner: { type: String, required: false, trim: true},
     content: { type: String, required: false },
     publishedIn: { type: Date, required: true },
     featuredImage: { type: String, required: false }, // Path to the featured image
@@ -31,6 +18,7 @@ const ArticleSchema: Schema<Article> = new Schema(
     trashed: { type: Boolean, default: false },
     category: { type: String, required: true },
     author: { type: String, required: true },
+    link: { type: String, required: true },
   },
   {
     timestamps: true, // Adds createdAt and updatedAt fields automatically
