+++ src/types/apiResoponse.ts
@@ -2,3 +2,4 @@ export interface ApiResponse {
     success: boolean;
     message: string;
 }
+
